Refactor common database utils into a module
